chore: move SlackMention into slack-notification.ts
